separate test case for base intensifier, flake fixes
